Rename project to kotlin-spring-boot-template
Updates root project name, subproject dir, Kotlin package
(com.yonatankarp.spring.starter -> com.yonatankarp.kotlin.spring.boot.template),
buildSrc convention plugin IDs, Docker/Postgres identifiers, CI matrix,
README badges, and init.py template placeholders.
