+++ buildSrc/src/main/kotlin/kotlin-spring-boot-template.java-conventions.gradle.kts
