+++ docker-compose.yml
@@ -1,6 +1,6 @@
 services:
   postgres:
-    container_name: spring-starter-postgres
+    container_name: kotlin-spring-boot-template-postgres
     networks:
       - proxynet
     image: postgres:18-alpine
@@ -8,7 +8,7 @@ services:
     environment:
       POSTGRES_USER: "postgres"
       POSTGRES_PASSWORD: "secret"
-      POSTGRES_DB: "spring-starter"
+      POSTGRES_DB: "kotlin-spring-boot-template"
     ports:
       - "5432:5432"
     expose:
@@ -16,4 +16,4 @@ services:
 
 networks:
   proxynet:
-    name: spring-starter-network
+    name: kotlin-spring-boot-template-network
