+++ kotlin-spring-boot-template/src/main/kotlin/com/yonatankarp/kotlin/spring/boot/template/Application.kt
@@ -1,0 +1,13 @@
+package com.yonatankarp.kotlin.spring.boot.template
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+@ConfigurationPropertiesScan
+class Application
+
+fun main(args: Array<String>) {
+    runApplication<Application>(*args)
+}
