+++ kotlin-spring-boot-template/src/main/kotlin/com/yonatankarp/kotlin/spring/boot/template/application/ports/input/Greet.kt
@@ -1,0 +1,7 @@
+package com.yonatankarp.kotlin.spring.boot.template.application.ports.input
+
+import com.yonatankarp.kotlin.spring.boot.template.domain.valueobject.Greeting
+
+fun interface Greet {
+    suspend operator fun invoke(): Greeting
+}
