+++ kotlin-spring-boot-template/src/main/kotlin/com/yonatankarp/kotlin/spring/boot/template/application/ports/output/GreetingCatalog.kt
@@ -1,0 +1,7 @@
+package com.yonatankarp.kotlin.spring.boot.template.application.ports.output
+
+import com.yonatankarp.kotlin.spring.boot.template.domain.valueobject.Greeting
+
+fun interface GreetingCatalog {
+    suspend fun random(): Greeting?
+}
