+++ kotlin-spring-boot-template/src/main/kotlin/com/yonatankarp/kotlin/spring/boot/template/domain/event/GreetingDelivered.kt
@@ -1,0 +1,9 @@
+package com.yonatankarp.kotlin.spring.boot.template.domain.event
+
+import com.yonatankarp.kotlin.spring.boot.template.domain.valueobject.Greeting
+import java.time.Instant
+
+data class GreetingDelivered(
+    val greeting: Greeting,
+    val deliveredAt: Instant,
+)
