+++ kotlin-spring-boot-template/src/main/resources/api/greetings-api.yaml
