+++ kotlin-spring-boot-template/src/main/resources/api/templates/kotlin-spring/apiInterface.mustache
