+++ kotlin-spring-boot-template/src/main/resources/logback-spring.xml
