+++ kotlin-spring-boot-template/src/main/resources/logback-stdout-appender.xml
