+++ kotlin-spring-boot-template/src/test/kotlin/com/yonatankarp/kotlin/spring/boot/template/ApplicationTest.kt
@@ -1,0 +1,24 @@
+package com.yonatankarp.kotlin.spring.boot.template
+
+import com.yonatankarp.kotlin.spring.boot.template.adapters.AbstractIntegrationTest
+import org.jooq.DSLContext
+import org.junit.jupiter.api.Assertions.assertNotNull
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.TestConstructor
+import org.springframework.test.context.TestConstructor.AutowireMode.ALL
+
+@SpringBootTest
+@TestConstructor(autowireMode = ALL)
+class ApplicationTest(
+    private val jooq: DSLContext,
+) : AbstractIntegrationTest() {
+    @Test
+    fun `smoke test`() {
+        assertNotNull(jooq)
+    }
+
+    @Test
+    fun `context loads`() {
+    }
+}
