--- spring-starter/src/main/kotlin/com/yonatankarp/spring/starter/Application.kt
@@ -1,13 +1,0 @@
-package com.yonatankarp.spring.starter
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-@ConfigurationPropertiesScan
-class Application
-
-fun main(args: Array<String>) {
-    runApplication<Application>(*args)
-}
