--- spring-starter/src/main/kotlin/com/yonatankarp/spring/starter/application/ports/input/Greet.kt
@@ -1,7 +1,0 @@
-package com.yonatankarp.spring.starter.application.ports.input
-
-import com.yonatankarp.spring.starter.domain.valueobject.Greeting
-
-fun interface Greet {
-    suspend operator fun invoke(): Greeting
-}
