--- spring-starter/src/main/kotlin/com/yonatankarp/spring/starter/application/ports/output/GreetingCatalog.kt
@@ -1,7 +1,0 @@
-package com.yonatankarp.spring.starter.application.ports.output
-
-import com.yonatankarp.spring.starter.domain.valueobject.Greeting
-
-fun interface GreetingCatalog {
-    suspend fun random(): Greeting?
-}
